@@ -1,8 +1,14 @@
 module github.com/antonyho/crhk-recorder
 
-go 1.15
+go 1.23
 
 require (
 	github.com/stretchr/testify v1.6.1
 	github.com/ushis/m3u v0.0.0-20150127162843-94396b784733
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
